Update of sketchpad.html
Added a couple of id's for styling - Note: you can't set the height and/or width of the canvas in the sketch.css file, else the drawing function goes loopy and the dot races away from the pointer (mouse or finger) I don't know why and am probably not going to look into it.
Just remember to set the canvas dimensions in the html file - haven't tested injecting style attribute with JS either. Though that may be useful at a later date.