Added a separate styles file for the sketch app
There a style definition to prevent the user inadvertently selecting text in the immediate vacinity of the canvas.

There's a couple of styles applied to the canvas element. DO NOT set height / width in sketch.ccs, else the drawing function goes doo-lally-tat - basically, the drawn dot runs away from the pointer :oP - look into it if you like. Just set cnavas dims in the html file for now!